requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
black==24.4.2
pytest==8.2.2
pytest-asyncio==0.23.7
python-jose==3.3.0
sqlalchemy==2.0.30
python-decouple==3.8
psycopg2-binary==2.9.9
alembic==1.13.1